+++ resolwe_bio/docker_images/common/packages-manual/r-packages.sh
@@ -14,3 +14,9 @@ sed -e 's/^#.*$//g' -e '/^$/d' /var/cache/build/packages-r.txt | \
 sed -e 's/^#.*$//g' -e '/^$/d' /var/cache/build/packages-r-bioconductor.txt | \
     Rscript --slave --no-save --no-restore-history \
         -e "source('http://www.bioconductor.org/biocLite.R'); biocLite(readLines('stdin'))"
+
+# XXX: This is unverifiable and thus may compromise the whole image.
+# XXX: Use notary (https://github.com/ropenscilabs/notary) when ready.
+Rscript --slave --no-save --no-restore-history -e " \
+  install.packages('http://hartleys.github.io/QoRTs/QoRTs_STABLE.tar.gz', repos=NULL, type='source') \
+"